Remove the "self" pointer from the sample server's protocol
