--- sample-server/src/main/java/com/proofpoint/platform/sample/PersonWithSelf.java
@@ -1,47 +1,0 @@
-/*
- * Copyright 2010 Proofpoint, Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.proofpoint.platform.sample;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-import com.google.auto.value.AutoValue;
-
-import java.net.URI;
-
-@AutoValue
-public abstract class PersonWithSelf
-{
-    public static PersonWithSelf createPersonWithSelf(Person person, URI self)
-    {
-        return new AutoValue_PersonWithSelf(person, self);
-    }
-
-    abstract Person getPerson();
-
-    @JsonProperty
-    public String getEmail()
-    {
-        return getPerson().getEmail();
-    }
-
-    @JsonProperty
-    public String getName()
-    {
-        return getPerson().getName();
-    }
-
-    @JsonProperty
-    public abstract URI getSelf();
-}
