+++ sample-server/src/main/java/com/proofpoint/platform/sample/PersonsResource.java
@@ -23,15 +23,10 @@ import com.proofpoint.platform.sample.PersonStore.StoreEntry;
 import javax.ws.rs.GET;
 import javax.ws.rs.Path;
 import javax.ws.rs.Produces;
-import javax.ws.rs.core.Context;
 import javax.ws.rs.core.MediaType;
-import javax.ws.rs.core.Response;
-import javax.ws.rs.core.UriBuilder;
-import javax.ws.rs.core.UriInfo;
-import java.net.URI;
+import java.util.Collection;
 
 import static com.google.common.base.Preconditions.checkNotNull;
-import static com.proofpoint.platform.sample.PersonWithSelf.createPersonWithSelf;
 
 @Path("/v1/person")
 public class PersonsResource
@@ -48,13 +43,12 @@ public class PersonsResource
 
     @GET
     @Produces(MediaType.APPLICATION_JSON)
-    public Response listAll(@Context UriInfo uriInfo)
+    public Collection<Person> listAll()
     {
-        Builder<PersonWithSelf> builder = ImmutableList.builder();
+        Builder<Person> builder = ImmutableList.builder();
         for (StoreEntry entry : store.getAll()) {
-            URI self = UriBuilder.fromUri(uriInfo.getRequestUri()).path(entry.getId()).build();
-            builder.add(createPersonWithSelf(entry.getPerson(), self));
+            builder.add(entry.getPerson());
         }
-        return Response.ok(builder.build()).build();
+        return builder.build();
     }
 }
